+++ src/Relation/HasOne.ts
@@ -17,10 +17,16 @@ class HasOne<T extends Odin = Odin> extends Relation<T, "HasOne"> {
     super(model, relation, localKey, foreignKey, filter, caller);
   }
 
-  public load(query: DB<T> | Join<T>, relations: Relation.Relation[], filter?: (q: Filter) => Filter) {
+  public load(
+    query: DB<T> | Join<T>, relations: Relation.Relation[], withTrashed?: boolean, filter?: (q: Filter) => Filter
+  ) {
     const relation = this.relation;
     const name = this.as;
-    const filters = [this.filter];
+    const filters: Array<(q: Filter) => Filter> = [];
+
+    if (relation.softDelete && !withTrashed) filters.push(q => q.whereNull(relation.DELETED_AT));
+
+    filters.push(this.filter);
 
     if (filter) filters.push(filter);
 
@@ -33,9 +39,9 @@ class HasOne<T extends Odin = Odin> extends Relation<T, "HasOne"> {
           if (!(relation as any)._relations.includes(subRelation))
             throw new Error(`Relation '${subRelation}' does not exist on '${relation.name}' Model`);
 
-          const loader = relation.prototype[subRelation]();
+          const loader: Relation = relation.prototype[subRelation]();
 
-          return loader.load(prev, cur.relations);
+          return loader.load(prev, cur.relations, withTrashed) as any;
         },
         filters.reduce(
           (prev, filter) => filter(prev) as any,
@@ -49,9 +55,16 @@ class HasOne<T extends Odin = Odin> extends Relation<T, "HasOne"> {
     });
   }
 
-  public loadCount(query: DB<T> | Join<T>, relations: string[], filter?: (q: Filter) => Filter) {
+  public loadCount(
+    query: DB<T> | Join<T>, relations: string[], withTrashed?: boolean, filter?: (q: Filter) => Filter
+  ) {
     const relation = this.relation;
     const subRelation = relations.shift();
+    const filters: Array<(q: Filter) => Filter> = [];
+
+    if (relation.softDelete && !withTrashed) filters.push(q => q.whereNull(relation.DELETED_AT));
+
+    filters.push(this.filter);
 
     if (subRelation) {
       if (!(relation as any)._relations.includes(subRelation))
@@ -60,25 +73,25 @@ class HasOne<T extends Odin = Odin> extends Relation<T, "HasOne"> {
       return query
         .join(
           relation.toString(),
-          q => relation.prototype[subRelation]().loadCount(
-            this.filter(
-              q
-                .where(this.foreignKey, `${this.model.constructor.toString()}.relation.${this.localKey}`)
-                .limit(1)
-                .aggregate({
-                  $project: {
-                    relation: "$$ROOT",
-                  },
-                })
-            ) as any,
-            relations,
-            filter
-          ),
+          q => relation.prototype[subRelation]()
+            .loadCount(
+              filters.reduce(
+                (prev, filter) => filter(prev) as any,
+                q
+                  .where(this.foreignKey, `${this.model.constructor.toString()}.relation.${this.localKey}`)
+                  .limit(1)
+                  .aggregate({
+                    $project: {
+                      relation: "$$ROOT",
+                    },
+                  })
+              ),
+              relations,
+              filter
+            ),
           "relation"
         );
     }
-
-    const filters = [this.filter];
 
     if (filter) filters.push(filter);